requirements.txt
Django==1.7.1
Pillow==2.6.1
django-bootstrap3==5.0.3
django-grappelli==2.6.3
django-markdown-deux==1.0.5
django-redis==3.8.0
markdown2==2.3.0
psycopg2==2.5.4
redis==2.10.3
sorl-thumbnail==12.1c
